// web-angular-nodejs/src/app/core/models/examineeResult.model.ts
export class ExamineeResultModel {
  userId: number = 0;
  firstName: string = '';
  lastName: string = '';
  email: string = '';
  objective: number = 0;
  subjective: number = 0;
  dateOfExam: string = '';
  objectiveResult: string = '';
  subjectiveResult: string = '';
  mailStatus: string = '';
  finalResult: string = '';
}